Refactor for target specific config: - Change environment to target for clarity
